requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
MarkupSafe==2.1.5
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==25.0
pillow==10.4.0
pluggy==1.5.0
pyautogui==0.9.54
PyGetWindow==0.0.9
pymsgbox==2.0.1
pyperclip==1.10.0
PyRect==0.2.0
pyscreeze==1.0.1
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-order==1.3.0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.4
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.27.0
trio-websocket==0.12.2
typing-extensions==4.13.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
